fed: Remove redundant e2e for secret and daemonset
Federation of daemonset and secret types is now implemented by the
sync controller, and e2e testing for each type is provided via crud
lifecycle e2e tests.  This renders the legacy e2e tests for these types
redundant, and this commit removes those tests.
