+++ test/e2e_federation/BUILD
@@ -13,13 +13,11 @@ go_library(
         "apiserver.go",
         "authn.go",
         "crud.go",
-        "daemonset.go",
         "deployment.go",
         "event.go",
         "ingress.go",
         "namespace.go",
         "replicaset.go",
-        "secret.go",
         "service.go",
         "upgrade.go",
         "util.go",
