+++ test/e2e_federation/ingress.go
@@ -32,6 +32,7 @@ import (
 	utilnet "k8s.io/apimachinery/pkg/util/net"
 	"k8s.io/apimachinery/pkg/util/wait"
 	fedclientset "k8s.io/kubernetes/federation/client/clientset_generated/federation_clientset"
+	"k8s.io/kubernetes/federation/pkg/federation-controller/util"
 	"k8s.io/kubernetes/pkg/api/v1"
 	"k8s.io/kubernetes/pkg/apis/extensions/v1beta1"
 	kubeclientset "k8s.io/kubernetes/pkg/client/clientset_generated/clientset"
@@ -52,6 +53,8 @@ const (
 	FederatedIngressTLSSecretName  = "federated-ingress-tls-secret"
 	FederatedIngressServicePodName = "federated-ingress-service-test-pod"
 	FederatedIngressHost           = "test-f8n.k8s.io."
+
+	FederatedSecretTimeout = 60 * time.Second
 
 	// TLS Certificate and Key for the ingress resource
 	// Generated using:
@@ -581,4 +584,55 @@ func getFederatedIngressAddress(client *fedclientset.Clientset, ns, name string)
 		}
 	}
 	return addresses, nil
+}
+
+func waitForSecretShardsOrFail(nsName string, secret *v1.Secret, clusters fedframework.ClusterSlice) {
+	framework.Logf("Waiting for secret %q in %d clusters", secret.Name, len(clusters))
+	for _, c := range clusters {
+		waitForSecretOrFail(c.Clientset, nsName, secret, true, FederatedSecretTimeout)
+	}
+}
+
+func waitForSecretOrFail(clientset *kubeclientset.Clientset, nsName string, secret *v1.Secret, present bool, timeout time.Duration) {
+	By(fmt.Sprintf("Fetching a federated secret shard of secret %q in namespace %q from cluster", secret.Name, nsName))
+	var clusterSecret *v1.Secret
+	err := wait.PollImmediate(framework.Poll, timeout, func() (bool, error) {
+		var err error
+		clusterSecret, err = clientset.Core().Secrets(nsName).Get(secret.Name, metav1.GetOptions{})
+		if (!present) && errors.IsNotFound(err) { // We want it gone, and it's gone.
+			By(fmt.Sprintf("Success: shard of federated secret %q in namespace %q in cluster is absent", secret.Name, nsName))
+			return true, nil // Success
+		}
+		if present && err == nil { // We want it present, and the Get succeeded, so we're all good.
+			By(fmt.Sprintf("Success: shard of federated secret %q in namespace %q in cluster is present", secret.Name, nsName))
+			return true, nil // Success
+		}
+		By(fmt.Sprintf("Secret %q in namespace %q in cluster.  Found: %v, waiting for Found: %v, trying again in %s (err=%v)", secret.Name, nsName, clusterSecret != nil && err == nil, present, framework.Poll, err))
+		return false, nil
+	})
+	framework.ExpectNoError(err, "Failed to verify secret %q in namespace %q in cluster: Present=%v", secret.Name, nsName, present)
+
+	if present && clusterSecret != nil {
+		Expect(util.SecretEquivalent(*clusterSecret, *secret))
+	}
+}
+
+func deleteSecretOrFail(clientset *fedclientset.Clientset, nsName string, secretName string, orphanDependents *bool) {
+	By(fmt.Sprintf("Deleting secret %q in namespace %q", secretName, nsName))
+	err := clientset.Core().Secrets(nsName).Delete(secretName, &metav1.DeleteOptions{OrphanDependents: orphanDependents})
+	if err != nil && !errors.IsNotFound(err) {
+		framework.ExpectNoError(err, "Error deleting secret %q in namespace %q", secretName, nsName)
+	}
+
+	// Wait for the secret to be deleted.
+	err = wait.Poll(5*time.Second, wait.ForeverTestTimeout, func() (bool, error) {
+		_, err := clientset.Core().Secrets(nsName).Get(secretName, metav1.GetOptions{})
+		if err != nil && errors.IsNotFound(err) {
+			return true, nil
+		}
+		return false, err
+	})
+	if err != nil {
+		framework.Failf("Error in deleting secret %s: %v", secretName, err)
+	}
 }
